requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
sentence-transformers
requests
beautifulsoup4
python-dotenv
